telegram_bridge/requirements.txt
python-telegram-bot==21.3
anthropic>=0.25.0
openai>=1.30.0
google-generativeai>=0.7.0
